requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
